Tile Adaptive and SI-UNIWARD embedding to bound memory on large covers
Both content-adaptive plugins built whole-image working sets (permutation,
cost map, STC trellis), so peak heap scaled with resolution and a ~13 MP
cover OOM'd on phones even with largeHeap. Partition the cover into
independent, memory-bounded bands and embed each as a self-contained
problem: per-band local permutation (password seed mixed with band index),
per-band cost, per-band STC, with the payload split deterministically across
bands by capacity. The header lives in band 0; extract recovers band geometry
from the image dimensions alone, so correctness is band-count-independent.

Band the JPEG codec too: the spatial sample planes and rounding-error side
information are recomputed per 8x8 block-row strip from a retained precover
reference instead of held full-resolution, and the resident quantized DCT
store is short[] rather than int[] (AC clamped to category 10, DC fits with
margin) — halving the dominant whole-image array. A 13.5 MP cover + 150 KB
payload now round-trips byte-identical under -Xmx256m (was ~900m); the new
:desktop:tilingTest enforces this floor so an O(image) regression fails loudly.

Backward compatibility for Adaptive/JpegUniward is intentionally not kept;
upstream LSB/RandomLSB and the OPENSTEGO format stamp are untouched.

Based on OpenStego by Samir Vaidya.

+++ desktop/src/test/java/com/openstego/desktop/image/jpeg/JpegCodecTest.java
@@ -139,7 +139,7 @@ class JpegCodecTest {
         int br = img.getBlocksHigh(comp) / 2;
         int bc = img.getBlocksWide(comp) / 2;
         int natIdx = 5; // an AC coefficient (natural-order index, not DC)
-        int[] before = img.getBlock(comp, br, bc).clone();
+        short[] before = img.getBlock(comp, br, bc).clone();
         img.getBlock(comp, br, bc)[natIdx] += 1;
 
         JpegImage round = JpegCodec.decode(JpegCodec.encode(img));
@@ -148,8 +148,8 @@ class JpegCodecTest {
         for (int c = 0; c < img.getComponentCount(); c++) {
             for (int r = 0; r < img.getBlocksHigh(c); r++) {
                 for (int cc = 0; cc < img.getBlocksWide(c); cc++) {
-                    int[] now = round.getBlock(c, r, cc);
-                    int[] want = img.getBlock(c, r, cc);
+                    short[] now = round.getBlock(c, r, cc);
+                    short[] want = img.getBlock(c, r, cc);
                     for (int k = 0; k < 64; k++) {
                         if (now[k] != want[k]) {
                             diffs++;
